#3 코드리뷰 반영 사항 APIResponseModel 형성 후 Model 분리
- 재사용 가능한 Response 모델을 따로 분리하여 파일 생성
- WeatherItem 주석추가
- Category를 enum으로 타입 지정
